+++ change_log.md
@@ -1,5 +1,10 @@
 ## Current Version
 
+### Version 1.7.11
+* Remove big endian check since exported blooms are identical on these systems
+* Set critical section to a named critical section
+* Fix divide by zero issue for Jaccard Index
+
 ### Version 1.7.10
 * Unique HashFunction type when using with related libraries